@@ -3,7 +3,7 @@ import { Store, select } from '@ngrx/store';
 import * as fromApp from '../../store/app.reducers';
 import * as AuthActions from '../../auth/store/auth.actions';
 import { User } from 'src/app/shared/user.model';
-import { FormBuilder } from '@angular/forms';
+import { UntypedFormBuilder } from '@angular/forms';
 import { Router, ActivatedRoute } from '@angular/router';
 
 @Component({
@@ -35,7 +35,7 @@ export class ProfileEditComponent implements OnInit {
 
   constructor(
     private store: Store<fromApp.AppState>,
-    private fb: FormBuilder,
+    private fb: UntypedFormBuilder,
     private router: Router,
     private route: ActivatedRoute
   ) {}
