@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { AttachmentComponent } from './attachment.component';
 
@@ -6,7 +6,7 @@ describe('AttachmentComponent', () => {
   let component: AttachmentComponent;
   let fixture: ComponentFixture<AttachmentComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ AttachmentComponent ]
     })
