@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { AttachmentsComponent } from './attachments.component';
 
@@ -6,7 +6,7 @@ describe('AttachmentsComponent', () => {
   let component: AttachmentsComponent;
   let fixture: ComponentFixture<AttachmentsComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ AttachmentsComponent ]
     })
