@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { QuestionShowComponent } from './question-show.component';
 
@@ -6,7 +6,7 @@ describe('QuestionShowComponent', () => {
   let component: QuestionShowComponent;
   let fixture: ComponentFixture<QuestionShowComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ QuestionShowComponent ]
     })
