@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { QuestionNavigateComponent } from './question-navigate.component';
 
@@ -6,7 +6,7 @@ describe('QuestionNavigateComponent', () => {
   let component: QuestionNavigateComponent;
   let fixture: ComponentFixture<QuestionNavigateComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ QuestionNavigateComponent ]
     })
