@@ -1,4 +1,4 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
 
 import { SpinerComponent } from './spiner.component';
 
@@ -6,7 +6,7 @@ describe('SpinerComponent', () => {
   let component: SpinerComponent;
   let fixture: ComponentFixture<SpinerComponent>;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     TestBed.configureTestingModule({
       declarations: [ SpinerComponent ]
     })
